refactor(J7VBGJ cross-scenario): extract HookResult + expectHookAllow/Deny to helpers.ts
Cross-scenario refactor pass for ticket J7VBGJ. The write-time and
commit-time gate integration tests each defined identical copies of
the HookResult interface, expectAllow(), and expectDeny() helpers
(~30 LOC of duplication). Extracted to packages/cli/tests/helpers.ts
as the single source of truth for any future PreToolUse hook test.

No behavior change: 8/8 tests still pass across both files.

(Originally marked skip:reason on the cross-scenario row claiming no
extraction emerged; on actually doing the sweep this duplication
surfaced. The skip-with-reason mechanism the feature ships is the
very thing that revealed the false skip — the user's "did we actually
run cross-scenario refactor?" prompt was the catch.)
